# optimize library and resource processing