--- a/app/src/main/java/org/ole/planet/myplanet/service/SyncManager.kt
+++ b/app/src/main/java/org/ole/planet/myplanet/service/SyncManager.kt
@@ -41,6 +41,8 @@ import java.io.IOException
 import java.util.Date
 import kotlin.system.measureTimeMillis
 import androidx.core.content.edit
+import org.ole.planet.myplanet.utilities.JsonUtils.getJsonObject
+import org.ole.planet.myplanet.utilities.SyncTimeLogger
 
 class SyncManager private constructor(private val context: Context) {
     private var td: Thread? = null
@@ -103,38 +105,110 @@ class SyncManager private constructor(private val context: Context) {
 
     private fun startFullSync() {
         try {
+            val logger = SyncTimeLogger.getInstance()
+            logger.startLogging()
+
             initializeSync()
 
             runBlocking {
                 val syncJobs = listOf(
-                    async { TransactionSyncManager.syncDb(mRealm, "tablet_users") },
-                    async { myLibraryTransactionSync() },
-                    async { TransactionSyncManager.syncDb(mRealm, "courses") },
-                    async { TransactionSyncManager.syncDb(mRealm, "exams") },
-                    async { TransactionSyncManager.syncDb(mRealm, "ratings") },
-                    async { TransactionSyncManager.syncDb(mRealm, "courses_progress") },
-                    async { TransactionSyncManager.syncDb(mRealm, "achievements") },
-                    async { TransactionSyncManager.syncDb(mRealm, "tags") },
-                    async { TransactionSyncManager.syncDb(mRealm, "submissions") },
-                    async { TransactionSyncManager.syncDb(mRealm, "news") },
-                    async { TransactionSyncManager.syncDb(mRealm, "feedback") },
-                    async { TransactionSyncManager.syncDb(mRealm, "teams") },
-                    async { TransactionSyncManager.syncDb(mRealm, "tasks") },
-                    async { TransactionSyncManager.syncDb(mRealm, "login_activities") },
-                    async { TransactionSyncManager.syncDb(mRealm, "meetups") },
-                    async { TransactionSyncManager.syncDb(mRealm, "health") },
-                    async { TransactionSyncManager.syncDb(mRealm, "certifications") },
-                    async { TransactionSyncManager.syncDb(mRealm, "team_activities") },
-                    async { TransactionSyncManager.syncDb(mRealm, "chat_history") }
+                    async {
+                        logger.startProcess("tablet_users_sync")
+                        TransactionSyncManager.syncDb(mRealm, "tablet_users")
+                        logger.endProcess("tablet_users_sync")
+                          },
+                    async {
+                        logger.startProcess("library_sync")
+                        myLibraryTransactionSync()
+                        logger.endProcess("library_sync")
+                          },
+                    async { logger.startProcess("courses_sync")
+                        TransactionSyncManager.syncDb(mRealm, "courses")
+                          logger.endProcess("courses_sync")
+                          },
+                    async { logger.startProcess("exams_sync")
+                        TransactionSyncManager.syncDb(mRealm, "exams")
+                          logger.endProcess("exams_sync")
+                          },
+                    async { logger.startProcess("ratings_sync")
+                        TransactionSyncManager.syncDb(mRealm, "ratings")
+                          logger.endProcess("ratings_sync")
+                          },
+                    async { logger.startProcess("courses_progress_sync")
+                        TransactionSyncManager.syncDb(mRealm, "courses_progress")
+                          logger.endProcess("courses_progress_sync")
+                          },
+                    async { logger.startProcess("achievements_sync")
+                        TransactionSyncManager.syncDb(mRealm, "achievements")
+                          logger.endProcess("achievements_sync")
+                          },
+                    async { logger.startProcess("tags_sync")
+                        TransactionSyncManager.syncDb(mRealm, "tags")
+                          logger.endProcess("tags_sync")
+                          },
+                    async { logger.startProcess("submissions_sync")
+                        TransactionSyncManager.syncDb(mRealm, "submissions")
+                          logger.endProcess("submissions_sync")
+                          },
+                    async { logger.startProcess("news_sync")
+                        TransactionSyncManager.syncDb(mRealm, "news")
+                          logger.endProcess("news_sync")
+                          },
+                    async { logger.startProcess("feedback_sync")
+                        TransactionSyncManager.syncDb(mRealm, "feedback")
+                          logger.endProcess("feedback_sync")
+                          },
+                    async { logger.startProcess("teams_sync")
+                        TransactionSyncManager.syncDb(mRealm, "teams")
+                          logger.endProcess("teams_sync")
+                          },
+                    async { logger.startProcess("tasks_sync")
+                        TransactionSyncManager.syncDb(mRealm, "tasks")
+                          logger.endProcess("tasks_sync")
+                          },
+                    async { logger.startProcess("login_activities_sync")
+                        TransactionSyncManager.syncDb(mRealm, "login_activities")
+                          logger.endProcess("login_activities_sync")
+                          },
+                    async { logger.startProcess("meetups_sync")
+                        TransactionSyncManager.syncDb(mRealm, "meetups")
+                          logger.endProcess("meetups_sync")
+                          },
+                    async { logger.startProcess("health_sync")
+                        TransactionSyncManager.syncDb(mRealm, "health")
+                          logger.endProcess("health_sync")
+                          },
+                    async { logger.startProcess("certifications_sync")
+                        TransactionSyncManager.syncDb(mRealm, "certifications")
+                          logger.endProcess("certifications_sync")
+                          },
+                    async { logger.startProcess("team_activities_sync")
+                        TransactionSyncManager.syncDb(mRealm, "team_activities")
+                          logger.endProcess("team_activities_sync")
+                          },
+                    async { logger.startProcess("chat_history_sync")
+                        TransactionSyncManager.syncDb(mRealm, "chat_history")
+                        logger.endProcess("chat_history_sync")
+                         }
                 )
 
                 syncJobs.awaitAll()
             }
 
+            logger.startProcess("admin_sync")
             ManagerSync.instance?.syncAdmin()
+            logger.endProcess("admin_sync")
+
+            logger.startProcess("resource_sync")
             resourceTransactionSync()
+            logger.endProcess("resource_sync")
+
+            logger.startProcess("on_synced")
             onSynced(mRealm, settings)
+            logger.endProcess("on_synced")
             mRealm.close()
+
+            logger.stopLogging()
         } catch (err: Exception) {
             err.printStackTrace()
             handleException(err.message)
@@ -285,17 +359,191 @@ class SyncManager private constructor(private val context: Context) {
     }
 
     private fun resourceTransactionSync(backgroundRealm: Realm? = null) {
-        val apiInterface = client?.create(ApiInterface::class.java)
+        val logger = SyncTimeLogger.getInstance()
+        logger.startProcess("resource_sync")
+
         try {
+            val apiInterface = client?.create(ApiInterface::class.java)
+
+            // Important: Execute Realm operations on the same thread
             if (backgroundRealm != null) {
-                syncResource(apiInterface, backgroundRealm)
+                // This is on a background thread, so we use the provided realm
+                syncResourceThreadSafe(apiInterface, backgroundRealm)
             } else {
-                syncResource(apiInterface)
+                // This is on the main thread with mRealm
+                syncResourceThreadSafe(apiInterface)
             }
         } catch (e: IOException) {
-            e.printStackTrace()
+            Log.e("SYNC", "Error in resourceTransactionSync: ${e.message}", e)
+        } finally {
+            logger.endProcess("resource_sync")
         }
     }
+
+    @Throws(IOException::class)
+    private fun syncResourceThreadSafe(dbClient: ApiInterface?, backgroundRealm: Realm? = null) {
+        val realmInstance = backgroundRealm ?: mRealm
+        val newIds: MutableList<String?> = ArrayList()
+
+        // Get all resource documents
+        val allDocs = dbClient?.getJsonObject(
+            Utilities.header,
+            "${Utilities.getUrl()}/resources/_all_docs?include_doc=false"
+        )
+        val all = allDocs?.execute()
+        val rows = getJsonArray("rows", all?.body())
+
+        // Collect all keys first
+        val allKeys = mutableListOf<String>()
+        for (i in 0 until rows.size()) {
+            val `object` = rows[i].asJsonObject
+            if (!TextUtils.isEmpty(getString("id", `object`))) {
+                allKeys.add(getString("key", `object`))
+            }
+        }
+
+        // Process in larger batches (2000 instead of 1000)
+        val batchSize = 2000
+        allKeys.chunked(batchSize).forEach { keys ->
+            if (keys.isNotEmpty()) {
+                val obj = JsonObject()
+                obj.add("keys", Gson().fromJson(Gson().toJson(keys), JsonArray::class.java))
+
+                val response = dbClient?.findDocs(
+                    Utilities.header,
+                    "application/json",
+                    "${Utilities.getUrl()}/resources/_all_docs?include_docs=true",
+                    obj
+                )?.execute()
+
+                if (response?.body() != null) {
+                    // Save is called on the same thread that created the Realm instance
+                    val ids: List<String?> = save(getJsonArray("rows", response.body()), realmInstance)
+                    newIds.addAll(ids)
+                }
+            }
+        }
+
+        // Remove deleted resources
+        removeDeletedResource(newIds, realmInstance)
+    }
+
+    private fun myLibraryTransactionSync(backgroundRealm: Realm? = null) {
+        val logger = SyncTimeLogger.getInstance()
+        logger.startProcess("library_sync")
+
+        try {
+            val apiInterface = client?.create(ApiInterface::class.java)
+
+            // Fetch all shelf documents
+            val res = apiInterface?.getDocuments(
+                Utilities.header,
+                "${Utilities.getUrl()}/shelf/_all_docs"
+            )?.execute()?.body()
+
+            val realmInstance = backgroundRealm ?: mRealm
+
+            // Process each shelf document sequentially to avoid Realm thread issues
+            for (i in res?.rows!!.indices) {
+                shelfDoc = res.rows!![i]
+                populateShelfItemsOptimized(apiInterface, realmInstance)
+            }
+        } catch (e: IOException) {
+            Log.e("SYNC", "Error in myLibraryTransactionSync: ${e.message}", e)
+        } finally {
+            logger.endProcess("library_sync")
+        }
+    }
+
+    private fun populateShelfItemsOptimized(apiInterface: ApiInterface?, realmInstance: Realm) {
+        try {
+            // Get the complete shelf document
+            val jsonDoc = apiInterface?.getJsonObject(
+                Utilities.header,
+                "${Utilities.getUrl()}/shelf/${shelfDoc?.id}"
+            )?.execute()?.body()
+
+            // Pre-fetch all documents that will be needed
+            val allIdsToFetch = mutableMapOf<String, MutableList<String>>()
+
+            // First pass - collect all IDs that need to be fetched, grouped by type
+            for (shelfData in Constants.shelfDataList) {
+                val array = getJsonArray(shelfData.key, jsonDoc)
+                if (array.size() > 0) {
+                    val ids = mutableListOf<String>()
+                    for (x in 0 until array.size()) {
+                        if (array[x] !is JsonNull) {
+                            ids.add(array[x].asString)
+                        }
+                    }
+                    if (ids.isNotEmpty()) {
+                        allIdsToFetch[shelfData.type] = ids
+                    }
+                }
+            }
+
+            // Second pass - fetch documents in batches for each type
+            for ((type, ids) in allIdsToFetch) {
+                // Process in batches of 100
+                val batchSize = 100
+                ids.chunked(batchSize).forEach { batch ->
+                    val obj = JsonObject()
+                    obj.add("keys", Gson().fromJson(Gson().toJson(batch), JsonArray::class.java))
+
+                    val response = apiInterface?.findDocs(
+                        Utilities.header,
+                        "application/json",
+                        "${Utilities.getUrl()}/$type/_all_docs?include_docs=true",
+                        obj
+                    )?.execute()
+
+                    if (response?.body() != null) {
+                        val rows = getJsonArray("rows", response.body())
+
+                        // Process all documents in one transaction
+                        realmInstance.executeTransaction { realm ->
+                            for (j in 0 until rows.size()) {
+                                val rowObj = rows[j].asJsonObject
+                                val doc = getJsonObject("doc", rowObj)
+
+                                // Set up string array for proper insertion
+                                stringArray[0] = shelfDoc?.id
+                                stringArray[1] = "" // Set appropriate category ID if needed
+                                stringArray[2] = type
+
+                                // Insert based on type
+                                when (type) {
+                                    "resources" -> insertMyLibrary(shelfDoc?.id, doc, realm)
+                                    "meetups" -> insert(realm, doc)
+                                    "courses" -> insertMyCourses(shelfDoc?.id, doc, realm)
+                                    "teams" -> insertMyTeams(doc, realm)
+                                }
+                            }
+                        }
+                    }
+                }
+            }
+
+            // Save concatenated links
+            saveConcatenatedLinksToPrefs()
+
+        } catch (e: Exception) {
+            Log.e("SYNC", "Error in populateShelfItemsOptimized: ${e.message}", e)
+        }
+    }
+
+//    private fun resourceTransactionSync(backgroundRealm: Realm? = null) {
+//        val apiInterface = client?.create(ApiInterface::class.java)
+//        try {
+//            if (backgroundRealm != null) {
+//                syncResource(apiInterface, backgroundRealm)
+//            } else {
+//                syncResource(apiInterface)
+//            }
+//        } catch (e: IOException) {
+//            e.printStackTrace()
+//        }
+//    }
 
     @Throws(IOException::class)
     private fun syncResource(dbClient: ApiInterface?, backgroundRealm: Realm? = null) {
@@ -322,22 +570,22 @@ class SyncManager private constructor(private val context: Context) {
         removeDeletedResource(newIds, realmInstance)
     }
 
-    private fun myLibraryTransactionSync(backgroundRealm: Realm? = null) {
-        val apiInterface = client?.create(ApiInterface::class.java)
-        try {
-            val res = apiInterface?.getDocuments(Utilities.header, "${Utilities.getUrl()}/shelf/_all_docs")?.execute()?.body()
-            for (i in res?.rows!!.indices) {
-                shelfDoc = res.rows!![i]
-                if (backgroundRealm != null) {
-                    populateShelfItems(apiInterface, backgroundRealm)
-                } else {
-                    populateShelfItems(apiInterface)
-                }
-            }
-        } catch (e: IOException) {
-            e.printStackTrace()
-        }
-    }
+//    private fun myLibraryTransactionSync(backgroundRealm: Realm? = null) {
+//        val apiInterface = client?.create(ApiInterface::class.java)
+//        try {
+//            val res = apiInterface?.getDocuments(Utilities.header, "${Utilities.getUrl()}/shelf/_all_docs")?.execute()?.body()
+//            for (i in res?.rows!!.indices) {
+//                shelfDoc = res.rows!![i]
+//                if (backgroundRealm != null) {
+//                    populateShelfItems(apiInterface, backgroundRealm)
+//                } else {
+//                    populateShelfItems(apiInterface)
+//                }
+//            }
+//        } catch (e: IOException) {
+//            e.printStackTrace()
+//        }
+//    }
 
     private fun populateShelfItems(apiInterface: ApiInterface, backgroundRealm: Realm? = null) {
         try {