@pavi38 2025-05-06
I  === SYNC TIME SUMMARY ===
              I  Total sync time: 1 min 23 sec (1m 23.194s)
              I  Individual process times:
              I  resource_sync                 :     54.06s ( 65%)
              I  library_sync                  :     28.92s ( 35%)
              I  admin_sync                    :       82ms (  0%)
             I  courses_sync                  :        9ms (  0%)
              I  courses_progress_sync         :        5ms (  0%)
              I  tablet_users_sync             :        4ms (  0%)
            I  exams_sync                    :        2ms (  0%)
              I  achievements_sync             :        2ms (  0%)
             I  ratings_sync                  :        2ms (  0%)
             I  on_synced                     :        2ms (  0%)
             I  teams_sync                    :        1ms (  0%)
              I  certifications_sync           :        1ms (  0%)
             I  tags_sync                     :        1ms (  0%)
              I  news_sync                     :        1ms (  0%)
              I  login_activities_sync         :        0ms (  0%)
              I  team_activities_sync          :        0ms (  0%)
            I  health_sync                   :        0ms (  0%)
             I  submissions_sync              :        0ms (  0%)
             I  feedback_sync                 :        0ms (  0%)
             I  meetups_sync                  :        0ms (  0%)
              I  tasks_sync                    :        0ms (  0%)
t              I  chat_history_sync             :        0ms (  0%)
2025-05-06 11:35:35.756  9319-9565  SyncTimeLogger          org.ole.planet.myplanet              I  =========================